% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_solar_pos.R
\name{calc_solar_pos}
\alias{calc_solar_pos}
\title{Calculate Solar Position for Building Centroids}
\usage{
calc_solar_pos(centroids, time, buildings)
}
\arguments{
\item{centroids}{sf object containing centroids of building geometries}

\item{time}{POSIXct time for which to calculate the solar position}

\item{buildings}{sf object containing building geometries}
}
\value{
sf object with solar azimuth and elevation added as attributes
}
\description{
Calculate Solar Position for Building Centroids
}
