% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_shadow_progression.R
\name{calculate_hourly_shadow_offset}
\alias{calculate_hourly_shadow_offset}
\title{Calculate Hourly Shadows}
\usage{
calculate_hourly_shadow_offset(solar_results)
}
\arguments{
\item{solar_results}{A list containing hourly solar data, where each element represents
solar data for a specific hour (e.g., solar elevation, azimuth, etc.).
This data should have been generated by a previous solar position calculation.}
}
\value{
A list of calculated shadows for each hour. Each element in the list corresponds
        to an hour from the input `solar_results` and contains the resulting shadow information
        calculated based on the solar data.
}
\description{
This function calculates the shadows for each hour based on the provided solar results.
It applies the `building_offset` function to each hourly solar data in the input list.
}
