% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_timestamps.R
\name{calculate_hourly_shadows}
\alias{calculate_hourly_shadows}
\title{Calculate hourly shadows}
\usage{
calculate_hourly_shadows(solar_results)
}
\arguments{
\item{solar_results}{A list containing hourly solar data.}
}
\value{
A list of calculated shadows for each hour.
}
\description{
Calculate hourly shadows
}
