% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shadow.R
\name{calculate_shadow_length}
\alias{calculate_shadow_length}
\title{Calculate the shadow length based on building height and solar elevation angle}
\usage{
calculate_shadow_length(measuredHeight, solar_elevation_angle)
}
\arguments{
\item{measuredHeight}{Numeric value representing the measured height of the building (in meters)}

\item{solar_elevation_angle}{Numeric value representing the solar elevation angle (in degrees)}
}
\value{
Numeric value representing the shadow length (in meters)
}
\description{
Calculate the shadow length based on building height and solar elevation angle
}
