% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_solar_pos.R
\name{calculate_solar}
\alias{calculate_solar}
\title{Calculate Solar Position for Given Time}
\usage{
calculate_solar(building_sf, time)
}
\arguments{
\item{building_sf}{sf object containing building geometries}

\item{time}{POSIXct. The time for which to calculate the solar position.}
}
\value{
sf object containing building geometries and solar positions
}
\description{
This function calculates the solar position for the provided building data and time.
}
