% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_timestamps.R
\name{calculate_solar_day}
\alias{calculate_solar_day}
\title{Calculate Solar Position for a Full Day}
\usage{
calculate_solar_day(building_sf, date)
}
\arguments{
\item{building_sf}{sf object containing building geometries}

\item{date}{Date object specifying the day for calculations}
}
\value{
List of sf objects containing solar positions for each hour
}
\description{
Calculate Solar Position for a Full Day
}
