% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD1_data_processing.R
\name{clean_polygon}
\alias{clean_polygon}
\title{Clean and process building geometries by merging parts and handling multipolygons}
\usage{
clean_polygon(building_sf)
}
\arguments{
\item{building_sf}{sf object containing building geometries with part IDs}
}
\value{
sf object with cleaned and processed geometries, where multipolygons are simplified and parts are merged
}
\description{
Clean and process building geometries by merging parts and handling multipolygons
}
