% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_files.R
\name{convert_to_2D}
\alias{convert_to_2D}
\title{Convert building data to 2D polygons}
\usage{
convert_to_2D(building_data, crs_code, tolerance = 0.1)
}
\arguments{
\item{building_data}{Data frame containing building data with 3D coordinates}

\item{crs_code}{Character string, the coordinate reference system (CRS) code to be used}

\item{tolerance}{Numeric value for geometry simplification. A higher tolerance value will result in more simplified geometries by removing more vertices.}
}
\value{
sf object containing 2D building polygons in the specified CRS
}
\description{
This function converts 3D building data to 2D polygons. It processes the geometry
by extracting the x and y coordinates from the 3D coordinates, ensuring that the polygons are closed,
and then creates a simple feature (sf) object with the specified coordinate reference system (CRS).
The function also simplifies the geometries to reduce complexity by removing unnecessary vertices.
The simplification is performed using the `st_simplify` function from the `sf` package.
}
\examples{
\dontrun{
# Example usage
building_data <- data.frame(
  geometry = c(
    "10.0 20.0 0.0 15.0 25.0 0.0 20.0 20.0 0.0 10.0 20.0 0.0",
    "30.0 40.0 0.0 35.0 45.0 0.0 40.0 40.0 0.0 30.0 40.0 0.0"
  ),
  stringsAsFactors = FALSE
)
crs_code <- "EPSG:25833"
polygons_2d <- convert_to_2D(building_data, crs_code)
print(polygons_2d)
}
}
