% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_timestamps.R
\name{create_hourly_shadow_polygons}
\alias{create_hourly_shadow_polygons}
\title{Create hourly shadow polygons}
\usage{
create_hourly_shadow_polygons(shadows, batch_size = 50)
}
\arguments{
\item{shadows}{A list of shadows for each hour.}

\item{batch_size}{integer number of buildings to process in each batch (default = 50)}
}
\value{
A named list of shadow polygons for each hour.
}
\description{
Create hourly shadow polygons
}
