% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_shadow_progression.R
\name{create_hourly_sunlight_areas}
\alias{create_hourly_sunlight_areas}
\title{Create hourly sunlight areas}
\usage{
create_hourly_sunlight_areas(shadow_polygons, building_sf)
}
\arguments{
\item{shadow_polygons}{A named list of shadow polygons for each hour.}

\item{building_sf}{An sf object containing building geometries.}
}
\value{
A named list of sunlight areas for each hour.
}
\description{
Create hourly sunlight areas
}
