% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shadow_geometry.R
\name{create_shadow_map}
\alias{create_shadow_map}
\title{Create an interactive map showing buildings, shadows, and sunlight areas}
\usage{
create_shadow_map(buildings, shadows, sunlight, time)
}
\arguments{
\item{buildings}{sf object containing buildings}

\item{shadows}{sf object containing shadow polygons}

\item{sunlight}{sf object containing sunlight areas}

\item{time}{POSIXct object representing the time}
}
\value{
leaflet map object
}
\description{
Create an interactive map showing buildings, shadows, and sunlight areas
}
