% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shadow_geometry.R
\name{create_shadow_polygons}
\alias{create_shadow_polygons}
\title{Create shadow polygons from buildings}
\usage{
create_shadow_polygons(buildings, batch_size)
}
\arguments{
\item{buildings}{sf object containing processed buildings}

\item{batch_size}{integer number of buildings to process in each batch}
}
\value{
sf object containing dissolved shadow polygons
}
\description{
Create shadow polygons from buildings
}
