% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shadow_geometry.R
\name{create_time_control}
\alias{create_time_control}
\title{Create a time control HTML element for the map}
\usage{
create_time_control(time)
}
\arguments{
\item{time}{POSIXct object representing the time}
}
\value{
character string containing HTML for the time control
}
\description{
Create a time control HTML element for the map
}
