% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shadow_geometry.R
\name{dissolve_shadow_polygons}
\alias{dissolve_shadow_polygons}
\title{Dissolve Shadow Polygons into a Single Geometry}
\usage{
dissolve_shadow_polygons(shadow_hulls)
}
\arguments{
\item{shadow_hulls}{An `sf` object containing shadow hull geometries. The `geometry` column
should include individual shadow polygons.}
}
\value{
An `sf` object with a single geometry representing the union of all shadow polygons.
}
\description{
This function dissolves multiple shadow hull polygons into a single unified geometry
by applying a union operation to all the shadow geometries in the input. It is useful for
combining overlapping or adjacent shadow areas into one boundary.
}
\details{
The function uses `st_union` to combine the geometries. If no valid geometries are present
  in the input, an error is raised.
}
\examples{
# Example usage:
dissolved_shadow <- dissolve_shadow_polygons(shadow_hulls_sf)
plot(dissolved_shadow$geometry)

}
