% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD1_data_processing.R
\name{extract_buildings}
\alias{extract_buildings}
\title{Extract building data from an XML file}
\usage{
extract_buildings(xml_file)
}
\arguments{
\item{xml_file}{Character string, path to the XML file containing building data}
}
\value{
Data frame containing extracted building data (building ID, part ID, geometry, height, and file name)
}
\description{
Extract building data from an XML file
}
