% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_crs.R
\name{extract_crs}
\alias{extract_crs}
\title{Extract CRS from a file or the first file in a directory (GML or XML)}
\usage{
extract_crs(path)
}
\arguments{
\item{path}{Path to the GML or XML file, or a directory containing such files}
}
\value{
The corresponding EPSG code, or NA if the CRS could not be determined
}
\description{
This function extracts the coordinate reference system (CRS) from a GML or XML file,
reading the `srsName` attribute from the `<gml:Envelope>` element. If a directory is provided,
it uses the first file in the directory.
}
