% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_example_data}
\alias{get_example_data}
\title{Get example building data included with the package}
\usage{
get_example_data()
}
\value{
sf object containing building geometries from the example dataset
}
\description{
Get example building data included with the package
}
