% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_building_data}
\alias{load_building_data}
\title{Load and process building data from XML files}
\usage{
load_building_data(file_path, recursive = FALSE)
}
\arguments{
\item{file_path}{Character string. Path to a XML file or directory containing XML files}

\item{recursive}{Logical. If TRUE and file_path is a directory, search recursively for XML files}
}
\value{
sf object containing building geometries
}
\description{
Load and process building data from XML files
}
