% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{run_interact_shadow_map}
\alias{run_interact_shadow_map}
\title{Run Interactive Shadow Map Application}
\usage{
run_interact_shadow_map(map_layer, date)
}
\arguments{
\item{map_layer}{A named list of sf objects for each hour, containing `buildings`, `shadows`, and `sunlight` geometries in the respective coordinate system.}

\item{date}{A character string representing the date (e.g., `"2025-02-18"`).}
}
\value{
Launches a Shiny interactive map application.
}
\description{
Starts a Shiny application to visualize shadow progression over a day.
It takes precomputed map layers containing buildings, shadows, and sunlight areas.
}
