% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_buildings}
\alias{visualize_buildings}
\title{Visualize building geometries with Leaflet}
\usage{
visualize_buildings(building_sf)
}
\arguments{
\item{building_sf}{sf object containing building geometries with height and building ID}
}
\value{
A leaflet map object showing building geometries
}
\description{
Visualize building geometries with Leaflet
}
