% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_solar_pos.R
\name{calc_centroids}
\alias{calc_centroids}
\title{Calculate Centroids for Building Geometries}
\usage{
calc_centroids(buildings)
}
\arguments{
\item{buildings}{sf object containing building geometries}
}
\value{
sf object with centroids of the building geometries
}
\description{
This function calculates the centroids for the provided building geometries.
}
