% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_building_centroid_and_sun_position.R
\name{calc_centroids}
\alias{calc_centroids}
\title{Calculate centroids of building geometries}
\usage{
calc_centroids(buildings)
}
\arguments{
\item{buildings}{sf object containing building geometries}
}
\value{
sf object containing the centroids of the building geometries
}
\description{
Calculate centroids of building geometries
}
