% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_building_centroid_and_sun_position.R
\name{calc_solar_pos}
\alias{calc_solar_pos}
\title{Calculate solar position for building centroids}
\usage{
calc_solar_pos(centroids, time, buildings)
}
\arguments{
\item{centroids}{sf object containing building centroids}

\item{time}{POSIXct object representing the time for solar position calculation}

\item{buildings}{sf object containing building geometries (to be updated with solar positions)}
}
\value{
sf object containing building geometries along with calculated solar positions (azimuth and elevation)
}
\description{
Calculate solar position for building centroids
}
