% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shadow.R
\name{calculate_all_shadows}
\alias{calculate_all_shadows}
\title{Calculate shadow geometry for each building}
\usage{
calculate_all_shadows(buildings)
}
\arguments{
\item{buildings}{sf object containing building geometries and attributes such as height, solar azimuth, and solar elevation}
}
\value{
sf object containing buildings with an additional column `shadow_geometry` representing the calculated shadow polygons for each building
}
\description{
Calculate shadow geometry for each building
}
