% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shadow.R
\name{calculate_shadow_points}
\alias{calculate_shadow_points}
\title{Calculate the shadow points of a building based on solar position and shadow length}
\usage{
calculate_shadow_points(building, solar_pos, shadow_length)
}
\arguments{
\item{building}{sf object containing the geometry of a building (POLYGON geometry)}

\item{solar_pos}{Numeric vector of length 2, where the first element is the solar azimuth (in degrees)
and the second element is the solar elevation (in degrees)}

\item{shadow_length}{Numeric value representing the shadow length (in meters), calculated based on the building's height and solar elevation angle}
}
\value{
sf object containing the shadow polygon geometry of the building
}
\description{
Calculate the shadow points of a building based on solar position and shadow length
}
