% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_shadow_progression.R
\name{calculate_solar_day}
\alias{calculate_solar_day}
\title{Calculate Solar Position for a Full Day}
\usage{
calculate_solar_day(building_sf, date)
}
\arguments{
\item{building_sf}{sf object containing building geometries. Each geometry should represent
a building or structure for which solar position will be calculated.}

\item{date}{Date object (class `Date`) specifying the day for the calculations.
The time zone will be set to UTC.}
}
\value{
A list of sf objects. Each element in the list corresponds to a specific hour of the day
        (e.g., "2025-02-24 - 12:00") and contains the solar positions for the buildings at that hour.
        Only solar positions with a positive solar elevation are included.
}
\description{
This function calculates the solar position for each hour of the specified day.
For each hour, the solar position is computed for the provided building geometries,
and only positions with a positive solar elevation are returned.
}
