% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD1_data_processing.R
\name{clean_polygon}
\alias{clean_polygon}
\title{Clean polygon geometries}
\usage{
clean_polygon(building_sf)
}
\arguments{
\item{building_sf}{sf object containing building polygons}
}
\value{
sf object with cleaned polygons
}
\description{
Clean polygon geometries
}
