% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD1_data_processing.R
\name{convert_to_2D}
\alias{convert_to_2D}
\title{Convert building data to 2D polygons}
\usage{
convert_to_2D(building_data, tolerance = 0.1)
}
\arguments{
\item{building_data}{Data frame containing building data}

\item{tolerance}{Numeric value for geometry simplification}
}
\value{
sf object containing 2D building polygons
}
\description{
Convert building data to 2D polygons
}
