% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_building_shadow_map}
\alias{create_building_shadow_map}
\title{Main function to process buildings and create an interactive shadow map}
\usage{
create_building_shadow_map(buildings, time, batch_size = 50)
}
\arguments{
\item{buildings}{sf object containing building geometries with shadow_geometry column}

\item{time}{POSIXct object representing the time for shadow calculation}

\item{batch_size}{integer number of buildings to process in each batch}
}
\value{
leaflet map object
}
\description{
Main function to process buildings and create an interactive shadow map
}
