% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_shadow_map}
\alias{create_shadow_map}
\title{Create a map showing buildings, shadows, and sunlight areas}
\usage{
create_shadow_map(buildings_union, shadows, sunlight, time)
}
\arguments{
\item{buildings_union}{sf object containing the merged building polygons}

\item{shadows}{sf object containing shadow polygons}

\item{sunlight}{sf object containing sunlight areas}

\item{time}{POSIXct object representing the time}
}
\value{
leaflet map object
}
\description{
Create a map showing buildings, shadows, and sunlight areas
}
