% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shadow_geometry.R
\name{create_sunlight_areas}
\alias{create_sunlight_areas}
\title{Create sunlight areas by subtracting buildings and shadows from a bounding box}
\usage{
create_sunlight_areas(buildings, shadows)
}
\arguments{
\item{buildings}{sf object containing buildings}

\item{shadows}{sf object containing shadow polygons}
}
\value{
sf object containing sunlight areas
}
\description{
Create sunlight areas by subtracting buildings and shadows from a bounding box
}
