% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_example_data}
\alias{get_example_data}
\title{Get example building data included with the package}
\usage{
get_example_data()
}
\value{
A simple feature (sf) object containing building geometries from the example dataset.
}
\description{
This function loads example building data included with the package.
}
\examples{
\dontrun{
# Get example building data
example_data <- get_example_data()
print(example_data)
}
}
