% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shadow_geometry.R
\name{process_buildings}
\alias{process_buildings}
\title{Process buildings and create shadow geometries}
\usage{
process_buildings(buildings)
}
\arguments{
\item{buildings}{sf object containing building geometries with a geometry column}
}
\value{
sf object containing processed buildings
}
\description{
Process buildings and create shadow geometries
}
